Add CSS styling
Changed the width of elements with both the classes, "avatar" and
"proportioned", to 300px and the height to auto. So that the selected
element stays in proportioned and it does not get distorted. Changed
the width of elements with both the classes, "avatar" and "distorted",
to 200px and the height to double the width at 400px, in order to make
the selected element distorted.